First command is being recived by server! Needed to set the messages to fixed length to test, all of this coding without testing was driving me crazy. Logging is setup per project, might need to change to whole solution at some point.

+++ TanksServer/TanksCommon/SharedObjects/GameStatus.cs
@@ -9,7 +9,7 @@ namespace TanksCommon.SharedObjects
     [Serializable]
     class GameStatus : IMessage
     {
-        public int Id { get => 1; }
+        public short Id { get => 1; }
         public int GameId { get; set; }
         public int PlayerId { get; set; }
         public int Status { get; set; }
