+++ TanksServer/TanksCommon/SharedObjects/InvalidMove.cs
@@ -9,7 +9,7 @@ namespace TanksCommon.SharedObjects
     [Serializable]
     class InvalidMove : IMessage
     {
-        public int Id { get => 2; }
+        public short Id { get => 2; }
         public int GameId { get; set; }
         public int PlayerId { get; set; }
         public int MoveId { get; set; }
