+++ TanksServer/TanksCommon/SharedObjects/MoveAccepted.cs
@@ -9,7 +9,7 @@ namespace TanksCommon.SharedObjects
     [Serializable]
     class MoveAccepted : IMessage
     {
-        public int Id { get => 5; }
+        public short Id { get => 5; }
         public int GameId { get; set; }
         public int PlayerId { get; set; }
         public int MoveId { get; set; }
