+++ TanksServer/TanksCommon/SharedObjects/RequestMove.cs
@@ -10,7 +10,7 @@ namespace TanksCommon.SharedObjects
     [Serializable]
     class RequestMove : IMessage
     {
-        public int Id { get => 6; }
+        public short Id { get => 6; }
         public int GameId { get; set; }
         public int PlayerId { get; set; }
         public int MoveId { get; set; }
